# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Real-Time Analytics Pipeline API",
    "version": "1.0.0",
    "description": "Enterprise-grade real-time analytics pipeline API. Supports event ingestion, analytics queries, dashboard management, and real-time WebSocket streaming.",
    "contact": {
      "name": "Engineering Team",
      "email": "engineering@example.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "https://api.example.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "example": "VALIDATION_ERROR" },
              "message": { "type": "string", "example": "Validation failed" },
              "details": { "type": "object", "example": {} }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      },
...
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array" },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "page_size": { "type": "integer" },
          "total_pages": { "type": "integer" },
          "has_next": { "type": "boolean" },
          "has_previous": { "type": "boolean" }
        }
      },
...
      "AggregatedMetricResponse": {
        "type": "object",
        "properties": {
          "metric_id": { "type": "string", "format": "uuid" },
          "metric_name": { "type": "string" },
          "window": { "type": "string", "enum": ["minute", "hourly", "daily"] },
          "window_start": { "type": "string", "format": "date-time" },
          "window_end": { "type": "string", "format": "date-time" },
          "value": { "type": "number" },
          "count": { "type": "integer" },
          "dimensions": { "type": "object" }
        }
      },
      "AggregationListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/PaginatedResponse" },
          {
            "type": "object",
            "properties": {
              "items": { "type": "array", "items": { "$ref": "#/components/schemas/AggregatedMetricResponse" } }
            }
          }
        ]
      },
      "RealtimeMetricsResponse": {
        "type": "object",
        "properties": {
          "time_range_seconds": { "type": "integer" },
          "total_events": { "type": "integer" },
          "events_per_second": { "type": "number" },
          "by_type": { "type": "object", "additionalProperties": { "type": "integer" } },
          "by_source": { "type": "object", "additionalProperties": { "type": "integer" } }
        }
      },
...
      "RetentionApplyResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string", "example": "completed" },
          "results": {
            "type": "object",
            "properties": {
              "raw_events_deleted": { "type": "integer" },
              "hourly_aggregations_deleted": { "type": "integer" },
              "daily_aggregations_deleted": { "type": "integer" }
            }
          }
        }
      },
      "RetentionSummaryResponse": {
        "type": "object",
        "properties": {
          "raw_events": { "type": "object", "properties": {
            "retention_days": { "type": "integer" },
            "total_records": { "type": "integer" },
            "oldest_record": { "type": "string", "format": "date-time" },
            "newest_record": { "type": "string", "format": "date-time" }
          }},
          "hourly_aggregations": { "type": "object" },
          "daily_aggregations": { "type": "object" }
        }
      },
...
      "SamplingStatsResponse": {
        "type": "object",
        "properties": {
          "current_sample_rate": { "type": "number" },
          "total_events_sampled": { "type": "integer" },
          "total_events_processed": { "type": "integer" },
          "is_high_volume_mode": { "type": "boolean" }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/analytics/aggregations/{metric_name}": {
      "get": {
        "operationId": "get_aggregations",
        "summary": "Get aggregated metrics",
        "description": "Retrieves aggregated metrics for a given metric name and time window.",
        "tags": ["Analytics"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "metric_name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "window", "in": "query", "schema": { "type": "string", "enum": ["minute", "hourly", "daily"], "default": "minute" } },
          { "name": "start", "in": "query", "required": true, "schema": { "type": "string", "format": "date-time" } },
          { "name": "end", "in": "query", "required": true, "schema": { "type": "string", "format": "date-time" } },
          { "name": "dimension_key", "in": "query", "schema": { "type": "string" } },
          { "name": "dimension_value", "in": "query", "schema": { "type": "string" } },
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "name": "page_size", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "Aggregated metrics",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AggregationListResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/aggregations/{metric_name}/latest": {
      "get": {
        "operationId": "get_latest_aggregation",
        "summary": "Get latest aggregation",
        "description": "Returns the most recent aggregation for a given metric name and window.",
        "tags": ["Analytics"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "metric_name", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "window", "in": "query", "schema": { "type": "string", "enum": ["minute", "hourly", "daily"], "default": "minute" } }
        ],
        "responses": {
          "200": {
            "description": "Latest aggregation",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AggregatedMetricResponse" }
              }
            }
          },
          "404": {
            "description": "No aggregation found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/counts/types": {
      "get": {
        "operationId": "get_event_type_counts",
        "summary": "Get event type counts",
        "description": "Returns event counts grouped by event type for a time range.",
        "tags": ["Analytics"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "start", "in": "query", "required": true, "schema": { "type": "string", "format": "date-time" } },
          { "name": "end", "in": "query", "required": true, "schema": { "type": "string", "format": "date-time" } }
        ],
        "responses": {
          "200": {
            "description": "Event type counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": { "type": "integer" },
                  "example": { "page_view": 1500, "click": 800, "signup": 45 }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/counts/sources": {
      "get": {
        "operationId": "get_source_counts",
        "summary": "Get source counts",
        "description": "Returns event counts grouped by source for a time range.",
        "tags": ["Analytics"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "start", "in": "query", "required": true, "schema": { "type": "string", "format": "date-time" } },
          { "name": "end", "in": "query", "required": true, "schema": { "type": "string", "format": "date-time" } }
        ],
        "responses": {
          "200": {
            "description": "Source counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": { "type": "integer" },
                  "example": { "web-app": 2000, "mobile-app": 500, "api-gateway": 300 }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/analytics/realtime": {
      "get": {
        "operationId": "get_realtime_metrics",
        "summary": "Get real-time metrics",
        "description": "Returns a snapshot of real-time metrics for the specified time window.",
        "tags": ["Analytics"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "window_seconds", "in": "query", "schema": { "type": "integer", "minimum": 10, "maximum": 3600, "default": 60 } }
        ],
        "responses": {
          "200": {
            "description": "Real-time metrics snapshot",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RealtimeMetricsResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/retention/apply": {
      "post": {
        "operationId": "apply_retention_policy",
        "summary": "Apply retention policy",
        "description": "Manually triggers the data retention policy to delete expired data.",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Retention policy applied",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RetentionApplyResponse" }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/retention/summary": {
      "get": {
        "operationId": "get_retention_summary",
        "summary": "Get retention summary",
        "description": "Returns a summary of current data retention status.",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Retention summary",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RetentionSummaryResponse" }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/sampling/stats": {
      "get": {
        "operationId": "get_sampling_stats",
        "summary": "Get sampling statistics",
        "description": "Returns current adaptive sampling statistics (admin only).",
        "tags": ["Admin"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Sampling statistics",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SamplingStatsResponse" }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}